net-base: Update params sent to StringPrtintf methods
An upstream change (CL:5855070) makes StringPrintf more strict when
matching parameters to format type.  libchrome added a patch to preserve
the old functionality.  In order to remove that patch, existing code
needs to be updated to work with the upstream change.

BUG=b:367200458
TEST=FEATURES=test emerge-guybrush chromeos-base/net-base

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6039817
Tested-by: Nathan Muggli <nmuggli@google.com>
Commit-Queue: Nathan Muggli <nmuggli@google.com>

+++ net-base/rtnl_message.cc
@@ -430,11 +430,13 @@ std::string RTNLMessage::NeighborStatus::ToString() const {
 // |lifetime| is unsigned. Printing as signed so that infinity (0xfffffff) get
 // printed as -1. Same below.
 std::string RTNLMessage::RdnssOption::ToString() const {
-  return base::StringPrintf("RdnssOption lifetime %d", lifetime);
+  return base::StringPrintf("RdnssOption lifetime %d",
+                            static_cast<int>(lifetime));
 }
 
 std::string RTNLMessage::DnsslOption::ToString() const {
-  return base::StringPrintf("DnsslOption lifetime %d", lifetime);
+  return base::StringPrintf("DnsslOption lifetime %d",
+                            static_cast<int>(lifetime));
 }
 
 std::string RTNLMessage::NdUserOption::ToString() const {
